requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==2.2.2
diff-match-patch==20121119
Django==2.0.5
django-cors-headers==2.2.0
django-crispy-forms==1.7.2
django-filter==1.1.0
django-formtools==2.1
django-import-export==1.0.0
django-pure-pagination==0.3.0
djangorestframework==3.8.2
et-xmlfile==1.0.1
future==0.15.2
httpie==0.9.9
httplib2==0.9.2
idna==2.7
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
Markdown==2.6.2
MarkupSafe==1.0
mysqlclient==1.3.12
odfpy==1.3.6
openpyxl==2.5.3
paramiko==2.4.1
Pillow==5.1.0
pyasn1==0.4.3
pycparser==2.18
Pygments==2.2.0
PyNaCl==1.2.1
pytz==2018.4
PyYAML==3.12
requests==2.19.1
six==1.10.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.23
xlrd==1.1.0
xlwt==1.3.0
